Inverted overlays order to fix small bug.
When the phone is located near a bus stop, the location marker goes on top of
the bus stop marker, which makes it impossible to access the stop times.

A possible workaround is to zoom until the stop marker appears again. I hope
this patch will fix the issue. The stop marker should now appear on top of the
current location marker.